Updated Scoring System in SpaceShip class...now with way less lines. Updated Initializing Debris.... now when we want Debris just simply ask for initDebris(). Updated Minion for Enemy... now you no longer have to make an MinionShooter for every enemy, just initMinion() in enemy's Constructor and minionMovement() in update method. Updated EnemyHud with a Healthbar ---> still needs the value of enemyhealth

+++ common/src/main/java/com/wisekrakr/firstgame/engine/gameobjects/weaponry/playerweaponry/MinionFighterPlayer.java
@@ -27,7 +27,7 @@ public class MinionFighterPlayer extends Minion {
     private static final float SPEED = 300;
     private float direction;
     private float radius;
-    private int health;
+    private float health;
 
     private float time;
     private int damage;
@@ -122,12 +122,12 @@ public class MinionFighterPlayer extends Minion {
     }
 
     @Override
-    public int getHealth() {
+    public float getHealth() {
         return health;
     }
 
     @Override
-    public void setHealth(int health) {
+    public void setHealth(float health) {
         this.health = health;
     }
