+++ common/src/main/java/com/wisekrakr/firstgame/engine/gameobjects/weaponry/playerweaponry/MinionShooterPlayer.java
@@ -28,7 +28,7 @@ public class MinionShooterPlayer extends Minion {
 
     private float direction;
     private float radius;
-    private int health;
+    private float health;
     private float shotLeftOver;
     private int ammoCount;
     private int damage;
@@ -143,12 +143,12 @@ public class MinionShooterPlayer extends Minion {
     }
 
     @Override
-    public int getHealth() {
+    public float getHealth() {
         return health;
     }
 
     @Override
-    public void setHealth(int health) {
+    public void setHealth(float health) {
         this.health = health;
     }
